create icon picker
